autenticacao basica metodos aluno

+++ Backend/AtividadePratica/AtividadePratica/Controllers/AlunoController.cs
@@ -23,12 +23,19 @@ public class AlunoController : ControllerBase
     [HttpPost]
     [Route("insert")]
     public async Task<ActionResult> InsertAsync(
-        Aluno? aluno)
+        [FromBody] Aluno? aluno)
     {
         try
         {
             if (aluno == null)
                 return new BadRequestObjectResult("Informe os dados do aluno");
+
+            var alunoComRu = await _alunoRepository.GetByRUAsync((int)aluno.RU);
+            
+            if (alunoComRu != null)
+            {
+                return new BadRequestObjectResult($"Aluno com RU {aluno.RU} já cadastrado!");
+            }
 
             //chama o repositorio para salvar o aluno
             var alunoSalvo = await _alunoRepository.InsertAsync(aluno);
@@ -42,17 +49,24 @@ public class AlunoController : ControllerBase
             return new BadRequestObjectResult(e.Message);
         }
     }
-    
+
     [HttpPut]
     [Route("update")]
     public async Task<ActionResult> UpdateAsync(
-        Aluno aluno)
+        [FromBody] Aluno aluno)
     {
         try
         {
             if (aluno == null)
                 return new BadRequestObjectResult("Informe os dados do aluno");
 
+            var alunoComRu = await _alunoRepository.GetByRUAsync((int)aluno.RU);
+            
+            if (alunoComRu == null)
+            {
+                return new NotFoundObjectResult($"Aluno com RU {aluno.RU} não encontrado!");
+            }
+            
             //chama o repositorio para atualizar o aluno
             var alunoSalvo = await _alunoRepository.UpdateAsync(aluno);
 
@@ -65,7 +79,7 @@ public class AlunoController : ControllerBase
             return new BadRequestObjectResult(e.Message);
         }
     }
-    
+
     [HttpGet]
     [Route("get")]
     public async Task<ActionResult> GetByRUAsync(
@@ -79,6 +93,11 @@ public class AlunoController : ControllerBase
             //chama o repositorio para carregar o aluno pelo RU
             var aluno = await _alunoRepository.GetByRUAsync((int)ru);
 
+            if (aluno == null)
+            {
+                return new NotFoundObjectResult($"Aluno com RU {ru} não encontrado!");
+            }
+            
             //retorna 200 com sucesso
             return new OkObjectResult(aluno);
         }
@@ -88,7 +107,7 @@ public class AlunoController : ControllerBase
             return new BadRequestObjectResult(e.Message);
         }
     }
-    
+
     [HttpDelete]
     [Route("delete")]
     public async Task<ActionResult> DeleteByRUAsync(
@@ -99,6 +118,13 @@ public class AlunoController : ControllerBase
             if (ru == null)
                 return new BadRequestObjectResult("Informe o RU");
 
+            var alunoComRu = await _alunoRepository.GetByRUAsync((int)ru);
+
+            if (alunoComRu == null)
+            {
+                return new NotFoundObjectResult($"Aluno com RU {ru} não encontrado!");
+            }
+            
             //chama o repositorio para deletar o aluno pelo RU
             await _alunoRepository.DeleteByRUAsync((int)ru);
 
@@ -111,7 +137,7 @@ public class AlunoController : ControllerBase
             return new BadRequestObjectResult(e.Message);
         }
     }
-    
+
     [HttpGet]
     [Route("get_by_curso")]
     public async Task<ActionResult> GetByCursoAsync(
